seconds_t double now

+++ timemanager/Clock.h
@@ -8,7 +8,7 @@
 #include <framework/globalfunctions/timevalOperations.h>
 
 typedef uint32_t millis_t;
-typedef float seconds_t;
+typedef double seconds_t;
 
 class Clock {
 public:
